+++ src/scanner/mod.rs
@@ -1,9 +1,12 @@
-mod container;
+mod ferox_scanner;
 mod utils;
 mod init;
 #[cfg(test)]
 mod tests;
+mod limit_heap;
+mod policy_data;
+mod requester;
 
-pub use self::container::{FeroxScanner, RESPONSES};
+pub use self::ferox_scanner::{FeroxScanner, RESPONSES};
 pub use self::init::initialize;
 pub use self::utils::PolicyTrigger;
